+++ packages/_infra/src/parameters.ts
@@ -1,5 +1,5 @@
 import { Stack } from '@aws-cdk/core';
-import * as AWS from 'aws-sdk';
+import CloudFormation from 'aws-sdk/clients/cloudformation';
 
 export const Parameters = {
     Edge: {
@@ -27,7 +27,7 @@ export interface ParametersEdge {
  * This function will lookup the configuration output from the `Edge` stack and provide them to following stacks
  */
 export async function getEdgeParameters(edge: Stack): Promise<null | ParametersEdge> {
-    const cfUsEast1 = new AWS.CloudFormation({ region: 'us-east-1' });
+    const cfUsEast1 = new CloudFormation({ region: 'us-east-1' });
     const edgeStack = await cfUsEast1.describeStacks({ StackName: edge.stackName }).promise();
     if (edgeStack == null) {
         console.log('Failed to lookup edge stack.. has it been deployed?');